sema: move `InitialValue` to separate file
